[refactor] all GPTModel included by `Model` struct
