+++ ChatBot/ChatBot/Models/GPTMessage/Implements/GPTMessage.swift
@@ -5,15 +5,27 @@
 //  Created by 김준성 on 1/2/24.
 //
 
-struct GPTMessage: Codable {
-    let role: GPTMessageRole
-    let content: String?
-    let name: String?
-    let toolCalls: [GPTToolCall]?
-    let toolCallID: String?
+extension Model {
+    struct GPTMessage: Codable, GPTMessagable {
+        let role: GPTMessageRole
+        let content: String?
+        let name: String?
+        let toolCalls: [GPTToolCall]?
+        let toolCallID: String?
+        
+        init(requestMessage: Model.GPTMessage) {
+            self.role = requestMessage.role
+            self.content = requestMessage.content
+            self.name = requestMessage.name
+            self.toolCalls = requestMessage.toolCalls
+            self.toolCallID = requestMessage.toolCallID
+        }
+        
+        func asRequestMessage() -> Model.GPTMessage { self }
+    }
 }
 
-extension GPTMessage {
+extension Model.GPTMessage {
     enum CodingKeys: String, CodingKey {
         case role
         case content
@@ -23,13 +35,15 @@ extension GPTMessage {
     }
 }
 
-struct GPTToolCall: Codable {
-    var id: String
-    var type: String
-    var function: Function
+extension Model {
+    struct GPTToolCall: Codable {
+        var id: String
+        var type: String
+        var function: Function
+    }
 }
 
-extension GPTToolCall {
+extension Model.GPTToolCall {
     struct Function: Codable {
         var name: String
         var arguments: String
