+++ ChatBot/ChatBot/Models/Model.swift
@@ -1,0 +1,8 @@
+//
+//  Model.swift
+//  ChatBot
+//
+//  Created by 김준성 on 1/4/24.
+//
+
+struct Model {}
